requirements.txt
haystack-ai==2.7.0
pypdf==5.3.0
python-docx==1.1.2
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.0
python-multipart==0.0.9